requirements.txt
discord.py
genshin
aiohttp